update requirements for game

+++ src/index.ts
@@ -10,15 +10,58 @@ import yargs from 'yargs'
 config();
 
 /** PARSE ARGUMENTS */
-const args = yargs(process.argv.slice(2)).argv as unknown as { twitch?: string, game: string, giphy?: string, giphyOffset?: string, twitter?: string, discord?: string, prompt?: string }
+const args = yargs(process.argv.slice(2)).argv as unknown as { 
+    twitch?: string, 
+    game: string, 
+    useGiphy?: string,
+    giphy?: string,
+    giphyOffset?: string, 
+    twitter?: string, 
+    onlyNotifyIfLive?: string,
+    discord?: string, 
+    mentionEveryone?: string, 
+    prompt?: string 
+}
 
-const twitchChannel = args["twitch"] as string || process.env.TWITCH_CHANNEL
+const twitchChannel = args["twitch"] ? args["twitch"] : process.env.TWITCH_CHANNEL
 const game = args["game"]
+if (game === undefined) {
+    console.error("please supply a --game")
+    process.exit()
+}
+const useGiphy = args['useGiphy'] ? args['useGiphy'] === "true" : true
 const giphySearch = args["giphy"] ? args["giphy"] : game
 const giphyOffset = args["giphyOffset"] ? args["giphyOffset"] : 0
 const twitterUser = args["twitter"] ? args["twitter"] : process.env.TWITTER_USER
+const onlyNotifyIfLive = args["onlyNotifyIfLive"] ? args["onlyNotifyIfLive"] === 'true' : true
 const discord = args["discord"] ? args["discord"] : process.env.DISCORD_CHANNEL_ID
-const prompt = args["prompt"] ? args["prompt"] : `Create a unique and expressive tweet with newlines for going live on twitch using ${twitchChannel} as the channel and ${game} as the game and https://twitch.tv/${twitchChannel} as the channel link. Must be less than 250 characters`
+const mentionEveryone = args["mentionEveryone"] ? args["mentionEveryone"] === 'true' : true
+
+const promptArguments = {
+    twitch: twitchChannel, 
+    game: game, 
+    useGiphy: useGiphy,
+    giphy: giphySearch,
+    giphyOffset: giphyOffset, 
+    twitter: twitterUser, 
+    discord: discord, 
+    mentionEveryone: mentionEveryone, 
+}
+
+
+const prompt = args["prompt"] ? (() => { 
+    let regexMatch = /\$\{([a-zA-Z0-9\-\_\|]+)\}/gm;
+    let matches = args["prompt"].match(regexMatch);
+    let prompt = args["prompt"];
+    if (matches !== null) {
+        matches.forEach(function (match) {           
+            let value = (promptArguments as any)[match.substr(2, match.length - 3)];
+            prompt = prompt.replace(match, value);
+        })
+    }
+    return prompt
+})() : `Create a unique and expressive tweet with newlines for going live on twitch using ${twitchChannel} as the channel and ${game} as the game and https://twitch.tv/${twitchChannel} as the channel link. Must be less than 250 characters`
+
 
 let twitchChannelData: any = undefined;
 
@@ -32,8 +75,7 @@ if (twitchChannel) {
     })).data.data[0]
 
     if (twitchChannelData === undefined) {
-        console.log(`${twitchChannel} is not live!`)
-        process.exit()
+        console.warn(`WARNING: ${twitchChannel} is not live!`)
     }
 } else {
     console.error('no --twitchChannel or TWITCH_CHANNEL found')
@@ -44,65 +86,90 @@ if (twitchChannel) {
  * POST TO TWITTER
  */
 if (twitterUser) {
-    try {
-        /** CONNECT CHATGPT */
-        const api = new ChatGPTAPI({
-            apiKey: process.env.OPENAI_API_KEY!,
-            completionParams: {
-                model: "gpt-4"
+    if (onlyNotifyIfLive && twitchChannelData !== undefined) {
+        try {
+            /** CONNECT CHATGPT */
+            const api = new ChatGPTAPI({
+                apiKey: process.env.OPENAI_API_KEY!,
+                completionParams: {
+                    model: "gpt-4"
+                }
+            })
+    
+            /** Ask ChatGPT to create the Tweet */
+            const chatgptTweet = await api.sendMessage(prompt);
+    
+            /** CONNECT TWITTER */
+            const twitterClient = new TwitterApi({
+                appKey: process.env.TWITTER_API_KEY,
+                appSecret: process.env.TWITTER_API_KEY_SECRET,
+                accessToken: process.env.TWITTER_ACCESS_TOKEN,
+                accessSecret: process.env.TWITTER_ACCESS_SECRET
+            } as TwitterApiTokens);
+    
+            /** Allow Twitter Client to Read and Write */
+            const rwTwitterClient = twitterClient.readWrite;
+    
+            let gifID: any = undefined
+    
+            if (useGiphy) {
+                /** UTLITY: Get the GIF from Giphy */
+                const getGIF = async (query: string, offset: string | number) => {
+                    const url = `https://api.giphy.com/v1/gifs/search?api_key=${process.env.GIPHY_API_KEY}&q=${query.split(" ").join("+")}&limit=25&offset=${offset}&rating=g&lang=en&bundle=messaging_non_clips`
+                    return (await axios.get(url)).data
+                }
+    
+                /** Search for the gif */
+                const gifSearch = await getGIF(giphySearch, giphyOffset)
+    
+                /** Fetch the gif */
+                const gif = (await axios.get('https://i.giphy.com/media/' + gifSearch.data[0].id + "/giphy.gif", {
+                    responseType: "text",
+                    responseEncoding: "base64",
+                })).data
+    
+                /** Download the gif */
+                fs.writeFileSync('./gif.gif', gif, { encoding: "base64" })
+    
+                /** Upload the gif to Twitter */
+                gifID = await rwTwitterClient.v1.uploadMedia('./gif.gif');
             }
-        })
+    
+            
+    
+            /** Create the tweet */
+    
+            let text = chatgptTweet.text.trim();
+            if (text.startsWith("\"")) {
+                text = text.substring(1)
+            }
 
-        /** Ask ChatGPT to create the Tweet */
-        const chatgptTweet = await api.sendMessage(prompt);
+            if (text.endsWith("\"")) {
+                text = text.substring(0, text.length - 2)
+            }
 
-        /** CONNECT TWITTER */
-        const twitterClient = new TwitterApi({
-            appKey: process.env.TWITTER_API_KEY,
-            appSecret: process.env.TWITTER_API_KEY_SECRET,
-            accessToken: process.env.TWITTER_ACCESS_TOKEN,
-            accessSecret: process.env.TWITTER_ACCESS_SECRET
-        } as TwitterApiTokens);
-
-        /** Allow Twitter Client to Read and Write */
-        const rwTwitterClient = twitterClient.readWrite;
-
-        /** UTLITY: Get the GIF from Giphy */
-        const getGIF = async (query: string, offset: string | number) => {
-            const url = `https://api.giphy.com/v1/gifs/search?api_key=uJKhh5fLXqJWJluguGo3e4bQUGYLop0o&q=${query.split(" ").join("+")}&limit=25&offset=${offset}&rating=g&lang=en&bundle=messaging_non_clips`
-            return (await axios.get(url)).data
+            console.log(text)
+    
+            let tweetPayload = {
+                text: text
+            } as SendTweetV2Params
+    
+            if (gifID) {
+                tweetPayload.media = {
+                    media_ids: [gifID]
+                }
+            }
+    
+            const tweetPostResult = await rwTwitterClient.v2.tweet(tweetPayload);
+    
+            console.log(`sent tweet https://twitter.com/${twitterUser}/status/${tweetPostResult.data.id}`)
+        } catch (error) {
+            console.error('failed to post to twitter');
+            console.error(error);
         }
-
-        /** Search for the gif */
-        const gifSearch = await getGIF(giphySearch, giphyOffset)
-
-        /** Fetch the gif */
-        const gif = (await axios.get('https://i.giphy.com/media/' + gifSearch.data[0].id + "/giphy.gif", {
-            responseType: "text",
-            responseEncoding: "base64",
-        })).data
-
-        /** Download the gif */
-        fs.writeFileSync('./gif.gif', gif, { encoding: "base64" })
-
-        /** Upload the gif to Twitter */
-        const id = await rwTwitterClient.v1.uploadMedia('./gif.gif');
-
-        /** Create the tweet */
-
-        const tweetPostResult = await rwTwitterClient.v2.tweet({
-            media: {
-                media_ids: [id]
-            },
-            text: chatgptTweet.text
-        } as SendTweetV2Params);
-
-        console.log(`sent tweet https://twitter.com/${twitterUser}/status/${tweetPostResult.data.id}`)
-    } catch (error) {
-        console.error('failed to post to twitter');
-        console.error(error);
+    } else {
+        console.log(`${twitchChannel} is not live, will not post tweet`)
     }
-
 } else {
     console.log(`no --twitter or TWITTER_USER found`)
 }
@@ -111,35 +178,42 @@ if (twitterUser) {
  * POST TO DISCORD
  */
 if (discord) {
-    try {
-        /** CONNECT DISCORD */
-        const client = new Client({
-            intents: [GatewayIntentBits.Guilds, GatewayIntentBits.MessageContent, GatewayIntentBits.GuildMessages]
-        });
-
-        /** DISCORD LOGIN */
-        await client.login(process.env.DISCORD_BOT_TOKEN!)
-
-        /** SEND TWITCH LIVE EMBED WHEN READY */
-        client.on('ready', async (client) => {
-            await client.channels.fetch(process.env.DISCORD_CHANNEL_ID!);
-            const embed = new EmbedBuilder();
-            embed.setTitle(`${twitchChannel} is now live on Twitch!`);
-            embed.setDescription(`**[Watch the stream](https://twitch.tv/${twitchChannel})**`);
-            embed.setColor(0x6441A4);
-            embed.setThumbnail(twitchChannelData.thumbnail_url);
-            embed.addFields(...[
-                { name: "Title", value: twitchChannelData.title } as APIEmbedField,
-                { name: "Viewers", value: twitchChannelData.viewer_count } as APIEmbedField
-            ]);
-            (client.channels.cache.get(process.env.DISCORD_CHANNEL_ID!) as TextChannel).send({ embeds: [embed] })
-            console.log(`sent message to discord`)
-
-            await client.destroy()
-        });
-    } catch (error) {
-        console.error("failed to post to discord")
-        console.error(error)
+    if (onlyNotifyIfLive && twitchChannelData !== undefined) {
+        try {
+            /** CONNECT DISCORD */
+            const client = new Client({
+                intents: [GatewayIntentBits.Guilds, GatewayIntentBits.MessageContent, GatewayIntentBits.GuildMessages]
+            });
+    
+            /** DISCORD LOGIN */
+            await client.login(process.env.DISCORD_BOT_TOKEN!)
+    
+            /** SEND TWITCH LIVE EMBED WHEN READY */
+            client.on('ready', async (client) => {
+                await client.channels.fetch(process.env.DISCORD_CHANNEL_ID!);
+                const embed = new EmbedBuilder();
+                embed.setTitle(`${twitchChannel} is${twitchChannelData ? ' now live ' : ' '}on Twitch!`);
+                embed.setDescription(`**[${twitchChannelData ? 'Watch the stream' : 'Check out the channel'}](https://twitch.tv/${twitchChannel})** ${mentionEveryone ? '@everyone' : ''}`);
+                embed.setColor(0x6441A4);
+                if (twitchChannelData) {
+                    embed.setThumbnail(twitchChannelData.thumbnail_url);
+                    embed.addFields(...[
+                        { name: "Game", value: game } as APIEmbedField,
+                        { name: "Title", value: twitchChannelData.title } as APIEmbedField,
+                        { name: "Viewers", value: twitchChannelData.viewer_count } as APIEmbedField
+                    ]);
+                }
+                (client.channels.cache.get(process.env.DISCORD_CHANNEL_ID!) as TextChannel).send({ embeds: [embed] })
+                console.log(`sent message to discord`)
+    
+                await client.destroy()
+            });
+        } catch (error) {
+            console.error("failed to post to discord")
+            console.error(error)
+        }
+    } else {
+        console.log(`${twitchChannel} is not live, will not post discord message`)
     }
 } else {
     console.log(`no --discord or DISCORD_CHANNEL_ID found`)
